use variable annotation and type checking

+++ julia/primesumsquares.jl
@@ -5,7 +5,7 @@
 
 """ Script to check an odd prime number has um of two squares """
 
-function sumsquares(nval)
+function sumsquares(nval::Int64)::Int64
     """An odd prime number p, the sum of
     two squares if and only if it leaves
     the remainder 1 on division by 4."""
@@ -16,9 +16,9 @@ function sumsquares(nval)
 end
 
 
-function prime(number)
+function prime(number::Int64)::Int64
     """function to check prime"""
-    sqrt_number = sqrt(number)
+    sqrt_number::Float64 = sqrt(number)
     for i in range(2, floor(Int64, sqrt_number))
         if number % i == 0
             return 0
@@ -29,7 +29,7 @@ end
 
 
 function main()
-    N = 100
+    N::Int64 = 100
     for n in range(2, N)
         prime(n)
     end
